requirements.txt
biopython==1.78
numpy==1.26.2
torch==1.10.1
tqdm==4.65.0
